% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/I-MDFA.R
\name{centraldev_original_func}
\alias{centraldev_original_func}
\title{This function transforms grand-mean parametrization back to original coefficients}
\usage{
centraldev_original_func(L, weight_h_exp)
}
\arguments{
\item{L}{Filter-length}

\item{weight_h_exp}{DFT of explanatory variables}
}
\value{
Q_centraldev_original Transformation matrix
}
\description{
This function transforms grand-mean parametrization back to original coefficients
}
