% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/I-MDFA.R
\name{des_mat_func}
\alias{des_mat_func}
\title{This function sets-up the data-based part of the design matrix, accounting for i1/i2 constraints (this is the effective design-matrix in the absence of regularization). It is originally written in grand-mean parametrization (which makes it more tedious).  Singularities in the case i1=F, i2=T are avoided by `displacing' the constraints by a small number.}
\usage{
des_mat_func(i2, i1, L, weight_h_exp, weight_constraint, shift_constraint, Lag)
}
\arguments{
\item{i2}{Boolean. If T a second-order filter constraint in frequency zero is obtained: time-shift of real-time filter must match target (together with i1 handles integration order two)}

\item{i1}{Boolean. If T a first-order filter constraint in frequency zero is obtained: amplitude of real-time filter must match weight_constraint (handles integration order one)}

\item{L}{Filter-length}

\item{weight_h_exp}{DFT of explanatory variables}

\item{weight_constraint}{Constraint vector in the case i1==T}

\item{shift_constraint}{Constraint vector in the case i2==T}

\item{Lag}{Nowcast (Lag=0), Forecast (Lag<0), Backcast (Lag>0)}
}
\value{
des_mat Data-based design matrix
}
\description{
This function sets-up the data-based part of the design matrix, accounting for i1/i2 constraints (this is the effective design-matrix in the absence of regularization). It is originally written in grand-mean parametrization (which makes it more tedious).  Singularities in the case i1=F, i2=T are avoided by `displacing' the constraints by a small number.
}
