% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DFA_code.R
\name{dfa_ms}
\alias{dfa_ms}
\title{This function computes MSE (mean-square) DFA-solutions: no customization, univariate problems only, no regularization, no constraints}
\usage{
dfa_ms(L, periodogram, Lag, Gamma)
}
\arguments{
\item{L}{Filter length}

\item{periodogram}{Periodogram (as generated by function per)}

\item{Lag}{Backcast (Lag>0), Nowcast (Lag=0) or Forecast (Lag<0)}

\item{Gamma}{Generic target specification: typically symmetric Lowpass (trend) or Bandpass (cycle) filters. Highpass and anticipative allpass (forecast) can be specified too.}
}
\value{
b optimal (MSE) filter coefficients

trffkt Complex transfer function of optimal (univariate MSE) filter
}
\description{
This function computes MSE (mean-square) DFA-solutions: no customization, univariate problems only, no regularization, no constraints
}
