% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/I-MDFA.R
\name{mat_func}
\alias{mat_func}
\title{This function sets-up the design matrix of the generic optimization problem: it combines data, regularization features and filter constraints}
\usage{
mat_func(i1, i2, L, weight_h_exp, lambda_decay, lambda_cross, lambda_smooth,
  Lag, weight_constraint, shift_constraint, grand_mean, b0_H0, c_eta, lag_mat)
}
\arguments{
\item{i1}{Boolean. If T a first-order filter constraint in frequency zero is obtained: amplitude of real-time filter must match weight_constraint (handles integration order one)}

\item{i2}{Boolean. If T a second-order filter constraint in frequency zero is obtained: time-shift of real-time filter must match target (together with i1 handles integration order two)}

\item{L}{Filter-length}

\item{weight_h_exp}{DFT of explanatory variables}

\item{lambda_decay}{Regularization: decay term}

\item{lambda_cross}{Regularization: cross-sectional term}

\item{lambda_smooth}{Regularization: smoothness term}

\item{Lag}{Nowcast (Lag=0), Forecast (Lag<0), Backcast (Lag>0)}

\item{weight_constraint}{Constraint vector in the case i1==T}

\item{shift_constraint}{Constraint vector in the case i2==T}

\item{grand_mean}{Boolean: if T then a grand-mean parametrization is imposed (default is F)}

\item{b0_H0}{Regularization: shrinkage target (arbitrary designs can be replicated by imposing strong regularization)}

\item{c_eta}{Boolean: impose mild/strong smoothness customization (default is F)}

\item{lag_mat}{Matrix for implementing effective lags in a mixed-frequency setting}
}
\value{
des_mat Design matrix

reg_mat Bilinear form (regularization matrix)

reg_xtxy Constants from regularization

w_eight Constants from filter constraints
}
\description{
This function sets-up the design matrix of the generic optimization problem: it combines data, regularization features and filter constraints
}
