% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/I-MDFA.R
\name{mdfa_analytic}
\alias{mdfa_analytic}
\title{Main estimation routine: Sets-up the generic optimization criteria proposed in MDFA-Legacy project (book)}
\usage{
mdfa_analytic(L, lambda, weight_func, Lag, Gamma, eta, cutoff, i1, i2,
  weight_constraint, lambda_cross, lambda_decay, lambda_smooth, lin_eta,
  shift_constraint, grand_mean, b0_H0, c_eta, weight_structure, white_noise,
  synchronicity, lag_mat, troikaner)
}
\arguments{
\item{L}{Filter-length}

\item{lambda}{Customization parameter: Timeliness is emphasized in the ATS-trilemma if lambda>0}

\item{weight_func}{DFT-matrix or alternative (for example model-based) estimate: first column is the target variable, additional columns are explanatory variables}

\item{Lag}{Nowcast (Lag=0), Forecast (Lag<0), Backcast (Lag>0)}

\item{Gamma}{Generic target specification: typically symmetric Lowpass (trend) or Bandpass (cycle) filters. Highpass and anticipative allpass (forecast) can be specified too}

\item{eta}{Customization parameter: Smoothness is emphasized in the ATS-trilemma if eta>0}

\item{cutoff}{Specifies start-frequency in stopband from which Smoothness is emphasized (corresponds typically to the cutoff of the lowpass target). Is not used if eta=0.}

\item{i1}{Boolean. If T a first-order filter constraint in frequency zero is obtained: amplitude of real-time filter must match weight_constraint (handles integration order one)}

\item{i2}{Boolean. If T a second-order filter constraint in frequency zero is obtained: time-shift of real-time filter must match target (together with i1 handles integration order two)}

\item{weight_constraint}{Constraint vector in the case i1==T}

\item{lambda_cross}{Regularization: cross-sectional term}

\item{lambda_decay}{Regularization: decay term}

\item{lambda_smooth}{Regularization: smoothness term}

\item{lin_eta}{Boolean: impose continuous or discontinuous Smoothness customization}

\item{shift_constraint}{Constraint vector in the case i2==T}

\item{grand_mean}{Boolean: if T then a grand-mean parametrization is imposed (default is F)}

\item{b0_H0}{Regularization: shrinkage target (arbitrary designs can be replicated by imposing strong regularization)}

\item{c_eta}{Boolean: impose mild/strong smoothness customization (default is F)}

\item{weight_structure}{Add structure to the optimization criterion (default value is weight_structure<-c(0,0): no structure imposed)}

\item{white_noise}{Impose a flat DFT (phase information is maintained)}

\item{synchronicity}{Impose a zero shift across series (amplitude information is maintained)}

\item{lag_mat}{Matrix for implementing effective lags in a mixed-frequency setting}

\item{troikaner}{Boolean: if T then degrees oif freedom will be computed (time-consuming computations if K is large)}
}
\value{
b Matrix of optimal filter coefficients

trffkt Complex transfer function of optimal multivariate filter

rever Criterion value (corresponds to a sample estimate of the MSE if lambda=eta=0)

degrees_freedom Degrees of freedom (when imposing regularization the degrres of freedom are smaller than L times the number of explanatory series)

Accuracy Accuracy term in decomposition of MSE

Smoothness Smoothness term in decomposition of MSE

Timeliness Timeliness term in decomposition of MSE

MS_error Sample estimate of MSE: consistent estimate in the cases lambda>0 and/or eta>0

freezed_degrees_new The complementary of degrees_freedom
}
\description{
Main estimation routine: Sets-up the generic optimization criteria proposed in MDFA-Legacy project (book)
}
