% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/I-MDFA.R
\name{MDFA_cust_constraint}
\alias{MDFA_cust_constraint}
\title{Simplified call for multivariate customization with filter constraints: no regularization}
\usage{
MDFA_cust_constraint(L, weight_func, Lag, Gamma, cutoff, lambda, eta, i1, i2,
  weight_constraint, shift_constraint)
}
\arguments{
\item{L}{Filter-length}

\item{weight_func}{DFT-matrix or alternative (for example model-based) estimate: first column is the target variable, additional columns are explanatory variables}

\item{Lag}{Nowcast (Lag=0), Forecast (Lag<0), Backcast (Lag>0)}

\item{Gamma}{Generic target specification: typically symmetric Lowpass (trend) or Bandpass (cycle) filters. Highpass and anticipative allpass (forecast) can be specified too}

\item{cutoff}{Specifies start-frequency in stopband from which Smoothness is emphasized (corresponds typically to the cutoff of the lowpass target). Is not used if eta=0.}

\item{lambda}{Customization parameter: Timeliness is emphasized in the ATS-trilemma if lambda>0}

\item{eta}{Customization parameter: Smoothness is emphasized in the ATS-trilemma if eta>0}

\item{i1}{Boolean. If T a first-order filter constraint in frequency zero is obtained: amplitude of real-time filter must match weight_constraint (handles integration order one)}

\item{i2}{Boolean. If T a second-order filter constraint in frequency zero is obtained: time-shift of real-time filter must match target (together with i1 handles integration order two)}

\item{weight_constraint}{Vector of amplitude constraints in frequency zero (typically 1 for lowpass and zero for bandpass or highpass)}

\item{shift_constraint}{Vector of time-shift constraints in frequency zero (typically zero for lowpass)}
}
\value{
mdfa_obj MDFA object
}
\description{
Simplified call for multivariate customization with filter constraints: no regularization
}

