% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/I-MDFA.R
\name{MS_decomp_total}
\alias{MS_decomp_total}
\title{This function decomposes the MSE-criterion into Accuracy, Smoothness and Timeliness components}
\usage{
MS_decomp_total(Gamma, trffkt, weight_func, cutoff, Lag)
}
\arguments{
\item{Gamma}{Generic target specification: typically symmetric Lowpass (trend) or Bandpass (cycle) filters. Highpass and anticipative allpass (forecast) can be specified too}

\item{trffkt}{Complex transfer function of optimal multivariate filter}

\item{weight_func}{DFT-matrix}

\item{cutoff}{Specifies start-frequency in stopband from which Smoothness is emphasized (corresponds typically to the cutoff of the lowpass target). Is not used if eta=0.}

\item{Lag}{Nowcast (Lag=0), Forecast (Lag<0), Backcast (Lag>0)}
}
\value{
Accuracy Accuracy term in decomposition of MSE

Smoothness Smoothness term in decomposition of MSE

Timeliness Timeliness term in decomposition of MSE

MS_error Sample estimate of MSE: consistent estimate in the cases lambda>0 and/or eta>0
}
\description{
This function decomposes the MSE-criterion into Accuracy, Smoothness and Timeliness components
}

