% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DFA_code.R
\name{per}
\alias{per}
\title{Simple function for computing DFT and periodogram (based on univariate DFA approach): cannot handle multivariate signal extraction}
\usage{
per(x, plot_T)
}
\arguments{
\item{x}{Data (time series)}

\item{plot_T}{Boolean (generate a plot of the periodogram: T/F)}
}
\value{
DFT Discrete Fourier Transform

per Periodogram (squared absolute DFT)
}
\description{
Simple function for computing DFT and periodogram (based on univariate DFA approach): cannot handle multivariate signal extraction
}

