% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DFT.r
\name{periodogram_bp}
\alias{periodogram_bp}
\title{Periodogram and DFT function: can handle more general cases than function per()}
\usage{
periodogram_bp(x, dd, n.pg)
}
\arguments{
\item{x}{Data}

\item{dd}{Integration order (default is d=0)}

\item{n.pg}{Length of time span considered for computing periodogram and DFT statistics}
}
\value{
perall Periodogram

fourtrans DFT
}
\description{
Periodogram and DFT function: can handle more general cases than function per()
}

