% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/I-MDFA.R
\name{Q_reg_func}
\alias{Q_reg_func}
\title{This function sets-up the regularization matrices}
\usage{
Q_reg_func(L, weight_h_exp, lambda_decay, lambda_smooth, lambda_cross, Lag,
  lag_mat, grand_mean, c_eta)
}
\arguments{
\item{L}{Filter-length}

\item{weight_h_exp}{DFT of explanatory variables}

\item{lambda_decay}{Regularization: decay term}

\item{lambda_smooth}{Regularization: smoothness term}

\item{lambda_cross}{Regularization: cross-sectional term}

\item{Lag}{Nowcast (Lag=0), Forecast (Lag<0), Backcast (Lag>0)}

\item{lag_mat}{Matrix for implementing effective lags in a mixed-frequency setting}

\item{grand_mean}{Boolean: if T then a grand-mean parametrization is imposed (default is F)}

\item{c_eta}{Boolean: impose mild/strong smoothness customization (default is F)}
}
\value{
Q_smooth Smoothness term (matrix)

Q_decay Decay term (matrix)

Q_cross Cross-sectional term (matrix)
}
\description{
This function sets-up the regularization matrices
}

