% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/I-MDFA.R
\name{reg_mat_func}
\alias{reg_mat_func}
\title{This function calls the regularization settings (bilinear forms) for cross-correlation, smoothness and decay patterns}
\usage{
reg_mat_func(weight_h_exp, L, c_eta, lambda_decay, Lag, grand_mean,
  lambda_cross, lambda_smooth, lag_mat)
}
\arguments{
\item{weight_h_exp}{DFT of explanatory variables}

\item{L}{Filter-length}

\item{c_eta}{Boolean: impose mild/strong smoothness customization (default is F)}

\item{lambda_decay}{Regularization: decay term}

\item{Lag}{Nowcast (Lag=0), Forecast (Lag<0), Backcast (Lag>0)}

\item{grand_mean}{Boolean: if T then a grand-mean parametrization is imposed (default is F)}

\item{lambda_cross}{Regularization: cross-sectional term}

\item{lambda_smooth}{Regularization: smoothness term}

\item{lag_mat}{Matrix for implementing effective lags in a mixed-frequency setting}
}
\value{
Q_smooth Smoothness term (matrix)

Q_cross Cross-sectional term (matrix)

Q_decay Decay term (matrix)

lambda_decay Re-parameterized decay-weight

lambda_smooth Re-parameterized smoothness-weight

lambda_cross Re-parameterized cross-sectional-weight
}
\description{
This function calls the regularization settings (bilinear forms) for cross-correlation, smoothness and decay patterns
}

