% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DFT.r
\name{spec_comp}
\alias{spec_comp}
\title{Generic function for computing the multivariate DFT: assigns a correct weight to frequency zero (1/sqrt(2)).}
\usage{
spec_comp(insamp, x, d)
}
\arguments{
\item{insamp}{Length of in-sample portion of data (by default insamp=nrow(x))}

\item{x}{(Multivariate) Data: first column is the target series, the remaining columns are the explanatory series.}

\item{d}{For d=1 a pseudo-DFT is implemented (default is d=0)}
}
\value{
weight_func (Multivariate) Discrete Fourier Transform
}
\description{
Generic function for computing the multivariate DFT: assigns a correct weight to frequency zero (1/sqrt(2)).
}

