% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/I-MDFA.R
\name{spec_mat_comp}
\alias{spec_mat_comp}
\title{This function sets-up the spectral matrix for closed-form solution of the optimization problem in the multivariate case: it relies on the multivariate DFT; it organizes lead/lags in the case of mixed-frequency data.}
\usage{
spec_mat_comp(weight_func, L, Lag, c_eta, lag_mat)
}
\arguments{
\item{weight_func}{DFT-matrix (first column is the traget variable, additional columns are explanatory variables)}

\item{L}{Filter-length}

\item{Lag}{Nowcast (Lag=0), Forecast (Lag<0), Backcast (Lag>0)}

\item{c_eta}{Boolean weight of frequency zero (default is F)}

\item{lag_mat}{Matrix for implementing effective lags in a mixed-frequency setting}
}
\value{
spec_mat Spectral matrix set-up for closed-form solution of optimization problem
}
\description{
This function sets-up the spectral matrix for closed-form solution of the optimization problem in the multivariate case: it relies on the multivariate DFT; it organizes lead/lags in the case of mixed-frequency data.
}

