% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/I-MDFA.R
\name{structure_func}
\alias{structure_func}
\title{This function allows for additional structure in the optimization criteria (features are not implemented yet)}
\usage{
structure_func(weight_func, spec_mat)
}
\arguments{
\item{weight_func}{DFT-matrix}

\item{spec_mat}{Spectral matrix}
}
\value{
Gamma_structure Target structure

spec_mat_structure Spectral matrix structure

Gamma_structure_diff Differenced target structure

spec_mat_structure Differenced spectral matrix structure
}
\description{
This function allows for additional structure in the optimization criteria (features are not implemented yet)
}

