% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/I-MDFA.R
\name{w_eight_func}
\alias{w_eight_func}
\title{This function specifies the constant in case of filter constraints with/without regularization}
\usage{
w_eight_func(i1, i2, Lag, weight_constraint, shift_constraint, L, weight_h_exp)
}
\arguments{
\item{i1}{Boolean. If T a first-order filter constraint in frequency zero is obtained: amplitude of real-time filter must match weight_constraint (handles integration order one)}

\item{i2}{Boolean. If T a second-order filter constraint in frequency zero is obtained: time-shift of real-time filter must match target (together with i1 handles integration order two)}

\item{Lag}{Nowcast (Lag=0), Forecast (Lag<0), Backcast (Lag>0)}

\item{weight_constraint}{Constraint vector in the case i1==T}

\item{shift_constraint}{Constraint vector in the case i2==T}

\item{L}{Filter-length}

\item{weight_h_exp}{DFT of explanatory variables}
}
\value{
w_eight Vector of filter weights entering closed-form solution
}
\description{
This function specifies the constant in case of filter constraints with/without regularization
}
