% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/I-MDFA.R
\name{white_noise_synchronicity}
\alias{white_noise_synchronicity}
\title{This function allows for particular parametrizations of DFT (not implemented yet)}
\usage{
white_noise_synchronicity(weight_func, white_noise, synchronicity)
}
\arguments{
\item{weight_func}{DFT-matrix}

\item{white_noise}{Boolean for flat spectrum}

\item{synchronicity}{Boolean for synchronous explanatory variables}
}
\value{
weight_func Parameterized DFT
}
\description{
This function allows for particular parametrizations of DFT (not implemented yet)
}

