% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/I-MDFA.R
\name{MDFA_MSE}
\alias{MDFA_MSE}
\title{Simplified call for multivariate MSE estimation: no regularization, no customization}
\usage{
MDFA_MSE(L, weight_func, Lag, Gamma, cutoff, i1, i2)
}
\arguments{
\item{L}{Filter-length}

\item{weight_func}{DFT-matrix}

\item{Lag}{Nowcast (Lag=0), Forecast (Lag<0), Backcast (Lag>0)}

\item{Gamma}{Generic target specification: typically symmetric Lowpass (trend) or Bandpass (cycle) filters. Highpass and anticipative allpass (forecast) can be specified too}

\item{cutoff}{Specifies start-frequency in stopband from which Smoothness is emphasized (corresponds typically to the cutoff of the lowpass target). Is not used if eta=0.}

\item{i1}{Boolean. If T a first-order filter constraint in frequency zero is obtained: amplitude of real-time filter must match weight_constraint (handles integration order one)}

\item{i2}{Boolean. If T a second-order filter constraint in frequency zero is obtained: time-shift of real-time filter must match target (together with i1 handles integration order two)}
}
\value{
mdfa_obj MDFA object
}
\description{
Simplified call for multivariate MSE estimation: no regularization, no customization
}

