% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/I-MDFA.R
\name{MDFA_reg}
\alias{MDFA_reg}
\title{Simplified call for multivariate MSE and customization with regularization but no constraints}
\usage{
MDFA_reg(L, weight_func, Lag, Gamma, cutoff, lambda, eta, lambda_cross,
  lambda_decay, lambda_smooth)
}
\arguments{
\item{L}{Filter-length}

\item{weight_func}{DFT-matrix or alternative (for example model-based) estimate: first column is the target variable, additional columns are explanatory variables}

\item{Lag}{Nowcast (Lag=0), Forecast (Lag<0), Backcast (Lag>0)}

\item{Gamma}{Generic target specification: typically symmetric Lowpass (trend) or Bandpass (cycle) filters. Highpass and anticipative allpass (forecast) can be specified too}

\item{cutoff}{Specifies start-frequency in stopband from which Smoothness is emphasized (corresponds typically to the cutoff of the lowpass target). Is not used if eta=0.}

\item{lambda}{Customization parameter: Timeliness is emphasized in the ATS-trilemma if lambda>0}

\item{eta}{Customization parameter: Smoothness is emphasized in the ATS-trilemma if eta>0}

\item{lambda_cross}{Regularization: cross-sectional term}

\item{lambda_decay}{Regularization: decay term}

\item{lambda_smooth}{Regularization: smoothness term}
}
\value{
mdfa_obj MDFA object
}
\description{
Simplified call for multivariate MSE and customization with regularization but no constraints
}
